Bug 1325406 - Part 2 - Fix 2 errors in last commit.
Unfortunately these changes did not have the desired effect of fixing the crash with optimizations enabled.   These were the only major changes visible in Seamonkey at the crash location which does not seem to experience this crash.  I will continue looking for the source of the crash.  Will debate merging these changes into master since they could be useful in the future.